Break dependency cycle by moving prompt context utils to dedicated file
