+++ front/lib/api/assistant/generation.test.ts
@@ -2,7 +2,7 @@ import { constructPromptMultiActions } from "@app/lib/api/assistant/generation";
 import {
   globalAgentInjectsUserContext,
   globalAgentInjectsWorkspaceContext,
-} from "@app/lib/api/assistant/global_agents/global_agents";
+} from "@app/lib/api/assistant/global_agents/prompt_context";
 import {
   normalizePrompt,
   systemPromptToText,
